Version check updater

+++ updater.ps1
@@ -3,6 +3,7 @@
 $URL = "https://github.com/BryanL43/VencordLargeUpload/releases/latest/download/dist.zip"
 $DownloadPath = "$env:APPDATA\Vencord\dist.zip"
 $ExtractPath = "$env:APPDATA\Vencord\dist"
+$ETagPath = "$env:APPDATA\Vencord\etag.txt"
 
 # Wait until internet is connected
 $Online = $false
@@ -19,18 +20,33 @@ while (-not $Online) {
 }
 
 try {
-    # Download the latest dist.zip
-    Invoke-WebRequest -Uri $URL -OutFile $DownloadPath -UseBasicParsing
+    # Get the ETag of the latest dist.zip
+    $Response = Invoke-WebRequest -Uri $URL -Method Head -UseBasicParsing
+    $ETag = $Response.Headers["ETag"]
+    $StoredETag = ""
 
-    # Unzip it into dist
-    if (Test-Path $ExtractPath) {
-        Remove-Item $ExtractPath -Recurse -Force
+    # Acquire stored ETag
+    if (Test-Path $ETagPath) {
+        $StoredETag = Get-Content $ETagPath -Raw
     }
-    Expand-Archive -LiteralPath $DownloadPath -DestinationPath $ExtractPath -Force
 
-    # Delete the downloaded dist.zip
-    if (Test-Path $DownloadPath) {
-        Remove-Item $DownloadPath -Force
+    if ($ETag -ne $StoredETag) {
+        # Download the latest dist.zip
+        Invoke-WebRequest -Uri $URL -OutFile $DownloadPath -UseBasicParsing
+
+        # Unzip it into dist
+        if (Test-Path $ExtractPath) {
+            Remove-Item $ExtractPath -Recurse -Force
+        }
+        Expand-Archive -LiteralPath $DownloadPath -DestinationPath $ExtractPath -Force
+
+        # Delete the downloaded dist.zip
+        if (Test-Path $DownloadPath) {
+            Remove-Item $DownloadPath -Force
+        }
+
+        # Save the new ETag
+        Set-Content -Path $ETagPath -Value $ETag
     }
 } catch {
     # Silently ignore errors so startup doesn't bother the user
